this version seems to be working (both for matsim repo and for freight-dfg17 repo).  ??
